backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
supabase==2.0.0
PyJWT==2.8.1
bcrypt==4.1.1
requests==2.31.0
Werkzeug==3.0.1
